+++ app/commands/characters_context/pathfinder2/feats/add_command.rb
@@ -5,20 +5,38 @@ module CharactersContext
     module Feats
       class AddCommand < BaseCommand
         use_contract do
+          Types = Dry::Types['strict.string'].enum('ancestry')
+
           params do
             required(:character).filled(type?: ::Pathfinder2::Character)
-            required(:feat).filled(type?: ::Pathfinder2::Feat)
+            required(:id).filled(:string)
+            required(:type).filled(Types)
+            required(:level).filled(:integer, gteq?: 1)
           end
         end
 
         private
 
-        def do_persist(input)
-          return { result: :ok } if ::Character::Feat.exists?(input)
+        def do_prepare(input)
+          input[:feat] = ::Pathfinder2::Feat.find(input[:id])
+        end
 
-          ::Character::Feat.create!(input.merge(ready_to_use: true))
+        def do_persist(input)
+          ::Character::Feat.create_with(ready_to_use: true).find_or_create_by(input.slice(:character, :feat))
+          update_selected_feats(input)
 
           { result: :ok }
+        end
+
+        def update_selected_feats(input)
+          selected_feats = input[:character].data.attributes['selected_feats'] || {}
+
+          values = selected_feats[input[:id]] || []
+          values << input.slice(:type, :level).stringify_keys
+          selected_feats[input[:id]] = values
+
+          input[:character].data = input[:character].data.attributes.merge('selected_feats' => selected_feats)
+          input[:character].save!
         end
       end
     end